ref(brain): Refactor `requiredChecks`...
Move logic out of `requiredChecks/index` and into different modules. The logic was getting a git out of control in there so separated into smaller functions.
