+++ src/brain/requiredChecks/constants.ts
@@ -1,0 +1,7 @@
+import { BuildStatus } from '@/config';
+
+export const OK_CONCLUSIONS = [
+  BuildStatus.SUCCESS,
+  BuildStatus.NEUTRAL,
+  BuildStatus.SKIPPED,
+] as string[];
